Implemented murmur hash 3.

+++ abstracts/Object.hh
@@ -10,7 +10,7 @@ public:
   virtual bool equals(Object& obj) const = 0;
 //State
 //  hash
-  virtual int hash() const = 0;
+  virtual unsigned long long hash() const = 0;
 //  clone
 //memory management may be up to the user.
 //clean up with delete
